+++ examples/queryWithParams.ts
@@ -9,7 +9,7 @@ import {
   flux,
   fluxDuration,
 } from '@influxdata/influxdb-client'
-import {url, token, org} from './env'
+import {url, token, org} from './env.mjs'
 
 const queryApi = new InfluxDB({url, token}).getQueryApi(org)
 const start = fluxDuration('-1m')