+++ examples/scripts/server.mjs
@@ -1,9 +1,9 @@
-const express = require('express')
-const path = require('path')
-const proxy = require('express-http-proxy')
-const open = require('open')
-const {url} = require('../env')
-const monitor = require('./monitor')
+import express from 'express'
+import proxy from 'express-http-proxy'
+import open from 'open'
+import {fileURLToPath} from 'node:url'
+import {url} from '../env.mjs'
+import monitor from './monitor.mjs'
 
 const port = 3001
 const proxyPath = '/influx'
@@ -12,7 +12,8 @@ const app = express()
 // monitor express response time in InfluxDB
 monitor(app)
 // serve all files of the git repository
-app.use(express.static(path.join(__dirname, '..', '..'), {index: false}))
+const dirName = fileURLToPath(new URL('../..', import.meta.url))
+app.use(express.static(dirName, {index: false}))
 // create also proxy to InfluxDB
 app.use(proxyPath, proxy(url))
 app.listen(port, () => {